# backend/docs/swagger.yaml
basePath: /api/
definitions:
...
  database.Lane:
    properties:
      competition_id:
        type: integer
      id:
        type: integer
      lane_number:
        type: integer
      players:
        items:
          $ref: '#/definitions/database.Player'
        type: array
      qualification_id:
        type: integer
    type: object
...
  database.Player:
    properties:
      group_id:
        type: integer
      id:
        type: integer
      lane_id:
        type: integer
      name:
        type: string
      order:
        type: integer
      participant_id:
        type: integer
      player_sets:
        items:
          $ref: '#/definitions/database.PlayerSet'
        type: array
      rank:
        type: integer
      rounds:
        items:
          $ref: '#/definitions/database.Round'
        type: array
      shoot_off_score:
        type: integer
      total_score:
        type: integer
    type: object
  database.PlayerSet:
    properties:
      elimination_id:
        type: integer
      id:
        type: integer
      players:
        items:
          $ref: '#/definitions/database.Player'
        type: array
      rank:
        type: integer
      set_name:
        type: string
      total_score:
        type: integer
    type: object
  database.Qualification:
    properties:
      advancing_num:
        type: integer
      end_lane:
        type: integer
      id:
        type: integer
      lanes:
        items:
          $ref: '#/definitions/database.Lane'
        type: array
      start_lane:
        type: integer
    type: object
  database.Round:
    properties:
      id:
        type: integer
      player_id:
        type: integer
      round_ends:
        items:
          $ref: '#/definitions/database.RoundEnd'
        type: array
      total_score:
        type: integer
    type: object
  database.RoundEnd:
    properties:
      id:
        type: integer
      is_confirmed:
        type: boolean
      round_id:
        type: integer
      round_scores:
        items:
          $ref: '#/definitions/database.RoundScore'
        type: array
    type: object
  database.RoundScore:
    properties:
      id:
        type: integer
      round_end_id:
        type: integer
      score:
        type: integer
    type: object
...
  endpoint.PutQualificationByID.QualificationPutData:
    properties:
      advancing_num:
        type: integer
      end_lane:
        type: integer
      start_lane:
        type: integer
    type: object
...
  response.ErrorIdResponse:
    properties:
      error:
        example: 'invalid ID(1) : sth error'
        type: string
    type: object
  response.ErrorInternalErrorResponse:
    properties:
      error:
        example: 'sth need fix ID(1) : sth error'
        type: string
    type: object
...
  response.Nill:
    type: object
...
host: localhost:80
info:
  contact:
    name: NYCUArchery
    url: https://github.com/NYCUarchery
  description: Gin swagger
  license:
    name: no license yet
  title: Gin swagger
  version: "1.0"
paths:
...
  /qualification/{id}:
    get:
      description: Get one Qualification without Lanes by id.
      parameters:
      - description: Qualification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success, return Qualification without lanes
          schema:
            allOf:
            - $ref: '#/definitions/database.Qualification'
            - properties:
                lanes:
                  $ref: '#/definitions/response.Nill'
              type: object
        "400":
          description: invalid qualification id
          schema:
            $ref: '#/definitions/response.ErrorIdResponse'
        "500":
          description: internall db error / Get Only Qualification
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Show one Qualification without Lanes.
      tags:
      - Qualification
    put:
      consumes:
      - application/json
      description: Put whole new Qualification and overwrite with the id, and update
        lanes below it ,but cannot replace groupid.
      parameters:
      - description: Qualification ID
        in: path
        name: id
        required: true
        type: integer
      - description: Qualification
        in: body
        name: Qualification
        required: true
        schema:
          $ref: '#/definitions/endpoint.PutQualificationByID.QualificationPutData'
      produces:
      - application/json
      responses:
        "200":
          description: success, return updated Qualification
          schema:
            $ref: '#/definitions/database.Qualification'
        "400":
          description: invalid qualification id / invalid lane id / lane is already
            occupied / invalid start or end lane number / Qualification is belong
            to UnassignedGroup, when update Qualification
          schema:
            $ref: '#/definitions/response.ErrorIdResponse'
        "500":
          description: internal db error / Update Lane Qualification Id / Update Qualification
            / Get Only Qualification
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Update one Qualification.
      tags:
      - Qualification
  /qualification/lanes/{id}:
    get:
      description: Get one Qualification with Lanes by id.
      parameters:
      - description: Qualification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success, return Qualification with lanes
          schema:
            allOf:
            - $ref: '#/definitions/database.Qualification'
            - properties:
                lanes:
                  allOf:
                  - $ref: '#/definitions/database.Lane'
                  - properties:
                      players:
                        $ref: '#/definitions/response.Nill'
                    type: object
              type: object
        "400":
          description: invalid qualification id
          schema:
            $ref: '#/definitions/response.ErrorIdResponse'
        "500":
          description: internall db error / Get Qualification With Lanes
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Show one Qualification with Lanes.
      tags:
      - Qualification
  /qualification/lanes/players/{id}:
    get:
      description: Get one Qualification with Lanes and Players by id
      parameters:
      - description: Qualification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success, return Qualification with lanes and players
          schema:
            allOf:
            - $ref: '#/definitions/database.Qualification'
            - properties:
                lanes:
                  allOf:
                  - $ref: '#/definitions/database.Lane'
                  - properties:
                      players:
                        allOf:
                        - $ref: '#/definitions/database.Player'
                        - properties:
                            player_sets:
                              $ref: '#/definitions/response.Nill'
                            rounds:
                              $ref: '#/definitions/response.Nill'
                          type: object
                    type: object
              type: object
        "400":
          description: invalid qualification id
          schema:
            $ref: '#/definitions/response.ErrorIdResponse'
        "500":
          description: internall db error / Get Qualification With Lanes and Players
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Show one Qualification
      tags:
      - Qualification
  /qualification/lanes/unassigned/{id}:
    get:
      description: Get one Qualification with Unassigned Lanes by id.
      parameters:
      - description: Qualification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success, return Qualification with Unassigned lanes
          schema:
            items:
              allOf:
              - $ref: '#/definitions/database.Qualification'
              - properties:
                  lanes:
                    allOf:
                    - $ref: '#/definitions/database.Lane'
                    - properties:
                        players:
                          allOf:
                          - $ref: '#/definitions/database.Player'
                          - properties:
                              player_sets:
                                $ref: '#/definitions/response.Nill'
                              rounds:
                                $ref: '#/definitions/response.Nill'
                            type: object
                      type: object
                type: object
            type: array
        "400":
          description: invalid qualification id
          schema:
            $ref: '#/definitions/response.ErrorIdResponse'
        "500":
          description: internall db error / Get Qualification With Lanes and Players
            / Get Qualification With Unassigned Lanes By ID
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Show one Qualification with Unassigned Lanes.
      tags:
      - Qualification
...
swagger: "2.0"
...